Add online mode fix for legacy versions; minor refactors in legacy
