+++ libraries/launcher/org/prismlauncher/utils/ReflectionUtils.java
@@ -54,15 +54,9 @@
 
 package org.prismlauncher.utils;
 
-import org.prismlauncher.utils.logging.Log;
-
-import java.applet.Applet;
-import java.io.File;
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
 
 public final class ReflectionUtils {
     private static final MethodHandles.Lookup LOOKUP = MethodHandles.lookup();
